+++ src/PipedOutput/Source/Helpers.cs
@@ -3,10 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using UnityEngine;
 
 namespace Nightinggale.PipedOutput
 {
-    public class Helpers
+    public static class Helpers
     {
         [System.Diagnostics.Conditional("DEBUG")]
         public static void PrintDebug(string text)
@@ -14,6 +15,12 @@ namespace Nightinggale.PipedOutput
             Console.Write(System.DateTime.UtcNow.ToString("[HH:mm:ss.fff]"));
             Console.Write(" [1] [DEBUG] [PipedOutput] ");
             Console.WriteLine(text);
+        }
+
+        public static void ManualDeliveryForce(this GameObject go)
+        {
+            foreach (var delivery in go.GetComponents<ManualDeliveryKG>())
+                delivery.operationalRequirement = FetchOrder2.OperationalRequirement.None;
         }
     }
 }