+++ src/PipedOutput/Changelog.md
@@ -1,5 +1,9 @@
 ﻿# Changelog
 
+## [1.4.4.0]
+- increased Oxygen Diffuser and RustDeoxidizer max pressure from to 4kg
+- added code to force manual delivery to RustDeoxidizer regardless of state
+
 ## [1.4.3.1]
 - removed logging